requirements.txt
# Core Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pandas>=2.1.1
python-dotenv>=1.0.0

# AI and OpenAI
openai>=1.3.0
tiktoken>=0.5.0

# Optional: Advanced AI Features (uncomment if needed)
# transformers>=4.36.0
# spacy>=3.7.2
# torch>=2.1.0

# Optional: Browser Automation (uncomment if needed)
# selenium>=4.15.0
# playwright>=1.40.0

# Optional: Async Support (uncomment if needed)
# aiohttp>=3.8.5
